+++ benches/contract/branch_elim/c/main.c
@@ -4,26 +4,27 @@
 // In BMB: contracts prove branch impossibility, enabling guaranteed elimination
 
 #include <stdio.h>
+#include <stdint.h>
 
 // Always returns positive (in BMB: post ret > 0)
-static long always_positive(long x) {
+static int64_t always_positive(int64_t x) {
     if (x > 0) return x;
     return 1;  // Ensures positive
 }
 
 // Always returns non-negative (in BMB: post ret >= 0)
-static long always_non_negative(long x) {
+static int64_t always_non_negative(int64_t x) {
     if (x >= 0) return x;
     return 0 - x;  // abs
 }
 
 // Process with unreachable negative branch
-static long process_positive(long n, long sum) {
+static int64_t process_positive(int64_t n, int64_t sum) {
     if (n <= 0) return sum;
 
-    long value = always_positive(n);
+    int64_t value = always_positive(n);
     // The negative branch below is dead code (value is always > 0)
-    long processed;
+    int64_t processed;
     if (value > 0) {
         processed = value * 2;
     } else {
@@ -34,12 +35,12 @@ static long process_positive(long n, long sum) {
 }
 
 // Process with multiple dead branches
-static long process_abs(long n, long sum) {
+static int64_t process_abs(int64_t n, int64_t sum) {
     if (n <= 0) return sum;
 
-    long value = always_non_negative(n);
+    int64_t value = always_non_negative(n);
     // Multiple dead branches (value is always >= 0)
-    long processed;
+    int64_t processed;
     if (value >= 0) {
         if (value > 1000000) {
             // Unlikely but possible
@@ -55,13 +56,13 @@ static long process_abs(long n, long sum) {
 }
 
 // Chained conditions with provable outcomes
-static long chained_positive(long x) {
-    long a = always_positive(x);
-    long b = always_positive(a);  // a > 0, so b > 0
-    long c = always_positive(b);  // b > 0, so c > 0
+static int64_t chained_positive(int64_t x) {
+    int64_t a = always_positive(x);
+    int64_t b = always_positive(a);  // a > 0, so b > 0
+    int64_t c = always_positive(b);  // b > 0, so c > 0
 
     // All these negative checks are dead branches
-    long result = 0;
+    int64_t result = 0;
     if (a > 0) result += a;
     if (b > 0) result += b;
     if (c > 0) result += c;
@@ -70,12 +71,12 @@ static long chained_positive(long x) {
 }
 
 // Benchmark: many operations with dead branches
-static long run_benchmark(long iterations, long sum) {
+static int64_t run_benchmark(int64_t iterations, int64_t sum) {
     if (iterations <= 0) return sum;
 
-    long r1 = process_positive(100, 0);
-    long r2 = process_abs(100, 0);
-    long r3 = chained_positive(iterations);
+    int64_t r1 = process_positive(100, 0);
+    int64_t r2 = process_abs(100, 0);
+    int64_t r3 = chained_positive(iterations);
 
     return run_benchmark(iterations - 1, sum + r1 + r2 + r3);
 }
@@ -83,18 +84,18 @@ static long run_benchmark(long iterations, long sum) {
 int main(void) {
     printf("Branch Elimination Benchmark\n\n");
 
-    long result1 = process_positive(100, 0);
-    printf("Process positive: %ld\n", result1);
+    int64_t result1 = process_positive(100, 0);
+    printf("Process positive: %lld\n", (long long)result1);
 
-    long result2 = process_abs(100, 0);
-    printf("Process abs: %ld\n", result2);
+    int64_t result2 = process_abs(100, 0);
+    printf("Process abs: %lld\n", (long long)result2);
 
-    long result3 = chained_positive(50);
-    printf("Chained positive: %ld\n", result3);
+    int64_t result3 = chained_positive(50);
+    printf("Chained positive: %lld\n", (long long)result3);
 
     // Benchmark
-    long bench = run_benchmark(100, 0);
-    printf("\nBenchmark (100 iterations): %ld\n", bench);
+    int64_t bench = run_benchmark(100, 0);
+    printf("\nBenchmark (100 iterations): %lld\n", (long long)bench);
 
     printf("Done.\n");
     return 0;
