+++ benches/contract/divzero_check/c/main.c
@@ -2,29 +2,30 @@
 // C has no contract system, must check manually or not at all
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
-static long safe_div(long a, long b) {
+static int64_t safe_div(int64_t a, int64_t b) {
     if (b == 0) { fprintf(stderr, "divzero\n"); exit(1); }
     return a / b;
 }
 
-static long safe_mod(long a, long b) {
+static int64_t safe_mod(int64_t a, int64_t b) {
     if (b == 0) { fprintf(stderr, "divzero\n"); exit(1); }
     return a % b;
 }
 
-static long div_chain(long x, long d) {
-    long a = safe_div(x, d);
-    long b = safe_mod(x, d);
-    long c = safe_div(a + b, d);
+static int64_t div_chain(int64_t x, int64_t d) {
+    int64_t a = safe_div(x, d);
+    int64_t b = safe_mod(x, d);
+    int64_t c = safe_div(a + b, d);
     return c;
 }
 
 int main(void) {
-    long acc = 0;
-    for (long n = 1000000; n > 0; n--) {
+    int64_t acc = 0;
+    for (int64_t n = 1000000; n > 0; n--) {
         acc += div_chain(n * 100 + 42, 7);
     }
-    printf("%ld\n", acc);
+    printf("%lld\n", (long long)acc);
     return 0;
 }
